feat: weather_screenのwidget test
