+++ test/unit/repository/weather_repository_impl_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.4 from annotations
-// in flutter_training/test/repository/weather_repository_impl_test.dart.
+// in flutter_training/test/unit/repository/weather_repository_impl_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
