+++ test/unit/ui/screen/weather/weather_screen_state_test.mocks.dart
@@ -1,11 +1,10 @@
 // Mocks generated by Mockito 5.4.4 from annotations
-// in flutter_training/test/application/weather_service_test.dart.
+// in flutter_training/test/unit/ui/screen/weather/weather_screen_state_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'package:flutter_training/application/weather_service.dart' as _i3;
 import 'package:flutter_training/domain/weather.dart' as _i2;
-import 'package:flutter_training/repository/weather_repository_impl.dart'
-    as _i3;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -31,42 +30,28 @@ class _FakeWeather_0 extends _i1.SmartFake implements _i2.Weather {
         );
 }
 
-/// A class which mocks [WeatherRepositoryImpl].
+/// A class which mocks [WeatherService].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockWeatherRepositoryImpl extends _i1.Mock
-    implements _i3.WeatherRepositoryImpl {
+class MockWeatherService extends _i1.Mock implements _i3.WeatherService {
   @override
-  _i2.Weather findBy(
-    String? area,
-    DateTime? date,
-  ) =>
-      (super.noSuchMethod(
+  _i2.Weather fetchWeather() => (super.noSuchMethod(
         Invocation.method(
-          #findBy,
-          [
-            area,
-            date,
-          ],
+          #fetchWeather,
+          [],
         ),
         returnValue: _FakeWeather_0(
           this,
           Invocation.method(
-            #findBy,
-            [
-              area,
-              date,
-            ],
+            #fetchWeather,
+            [],
           ),
         ),
         returnValueForMissingStub: _FakeWeather_0(
           this,
           Invocation.method(
-            #findBy,
-            [
-              area,
-              date,
-            ],
+            #fetchWeather,
+            [],
           ),
         ),
       ) as _i2.Weather);
